+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # master
 
+* Add Turntabler::Client#user_by_name for looking up users by their name instead of id
 * Allow authentication via user ids / auth tokens if a password is unavailable
 * Authenticate using emails / passwords instead of user ids / auth tokens
 * Fix exceptions in callbacks not causing one-time callbacks to be unregistered